Update readme with flamegraph example
Took me forever to actually figure out how the commands to generate an view a flamegraph were supposed to be used. I added an example.

Also wanted to give an example on how to set interval option in the same section where we're talking about default intervals.

Added a `#` before `...` so it's technically valid ruby code.

Uppercase section header.