Rewrite flamegraph viewer using canvas
Large profiles were generating so many SVG nodes that the browser was
bogging down. Using canvas keeps the DOM very simple, and repainting it
seems to be pretty fast. It also lets us drop the D3 and jQuery
dependencies for a smaller page load.

The viewer has been flipped vertically: the small overview graph is on
the top, and stacks grow top down instead of bottom up. This matches
Chrome's devtools timeline and many other timeline/flamegraph tools I've
seen.

Controls are:

 - drag: pan/scroll
 - scrollwheel: pan/scroll
 - Shift-scrollwheel: zoom

Click-and-drag to select a region in the small overview graph isn't
currently supported, nor is dragging the selected region to pan/scroll.
That would be nice to add back eventually, though I've found that
zoom/scroll performance is good enough that I haven't really missed
those features.

This should be compatible with IE 10 and newer, though I've only tested
in Safari, Firefox, and Chrome.