Minimize changes to canvas fill/stroke styles
We now first draw all frame backgrounds, then draw all frame labels. And
when drawing the backgrounds, we group them by color. This lets us set
the fillStyle a minimum number of times: once for each color, and once
before drawing all the labels.

This speeds up rendering of a 3MB profile from 38.5fps to 50fps in
Chrome, and from 50fps to 62.5fps in Safari.