Eliminate one loop over the stack by using `overall_samples`
Previously, the `already_accounted_in_total` member on the
`frame_data_t` struct was used to keep track of whether or not a
particular frame had been seen in a stack.  For example, a recursive
function is only supposed to count the frame once in the `total_samples`
member.

Instead of using the `already_accounted_in_total` as a flag, this commit
uses the `overall_samples` value as a unique identifier for the frame.
`overall_samples` will be different on every call to
`stackprof_record_sample_for_stack`, but the value will stay the same
for the duration of that function.  That means we can use the value to
indicate whether or not a particular frame has been "seen", and
we can eliminate one loop over the stack.