Record marking and sweeping as fake gc calls
This attempts to measure the states GC is in by recording fake frames of
"(marking)" or "(sweeping)". Occasonally we also encounter the GC being
in state :none, but this should be fairly rare and harmless (we record
this as a GC frame but not part of mark or sweep).

We don't have the ability to record the exact time mark and sweep
happens, so we just keep a counter of how many times we've encountered
each state. We assume all the marking happens before all the sweeping.