Remove reentrancy detection from postponed job
Because this is in the postponed job handler, we should never be able to
reenter this code.

I also believe this has a race condition if another signal arrives
between the if statement and the increment. Signals can also arrive in
any thread so this would not be a safe way to avoid reentrancy.