Make test_cputime more reliable
It seems the order of the frames hash values in test_cputime is somewhat
nondeterministic, at least with Ruby > 3. Most of the time the
"StackProfTest#math" frame will be at index 1, but sometimes it shows up
at index 0, like it does with Ruby < 3. This commit updates the test to
accept it in either position.